pyproject.toml
[project]
name = "sshkube"
version = "0.2.1"
description = "Access kubernetes clusters over ssh"
authors = [
    {name = "Daniel J. B. Clarke",email = "u8sand@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click (>=8.2.1,<9.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "winloop (>=0.1.8,<0.2.0) ; sys_platform == \"win32\""
]

[tool.poetry.scripts]
sshkube = 'sshkube.__main__:cli'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
